requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
idna==3.2
numpy==1.21.2
pandas==1.3.2
pydantic==1.8.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
sqlalchemy2-stubs==0.0.2a12
sqlmodel==0.0.4
starlette==0.14.2
todoist-python==8.1.3
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
